requirements.txt
torch>=2.0.0
torchvision>=0.15.0

numpy>=1.24.0
scipy>=1.10.0

# image
scikit-image>=0.21.0
Pillow>=9.5.0

# 3D mesh
trimesh>=3.22.0
rtree>=1.0.0

PyYAML>=6.0

tqdm>=4.65.0

matplotlib>=3.7.0

pytest>=7.3.0
